backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# AI/LLM - COMPATIBLE VERSIONS
openai==1.55.3
httpx==0.27.2
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain==0.3.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
requests==2.31.0
